src/main/java/org/eclipse/rdf4j/query/algebra/evaluation/function/postgis/geometry/relation/RelateMatch.java
package org.eclipse.rdf4j.query.algebra.evaluation.function.postgis.geometry.relation;

import org.eclipse.rdf4j.model.Value;
import org.eclipse.rdf4j.model.ValueFactory;
import org.eclipse.rdf4j.model.vocabulary.POSTGIS;
import org.eclipse.rdf4j.query.algebra.evaluation.ValueExprEvaluationException;
import org.eclipse.rdf4j.query.algebra.evaluation.function.Function;

public class RelateMatch implements Function {

	@Override
	public String getURI() {
		return POSTGIS.st_relateMatch.stringValue();
	}

	@Override
	public Value evaluate(ValueFactory valueFactory, Value... args) throws ValueExprEvaluationException {
		if (args.length != 2) {
			throw new ValueExprEvaluationException(getURI() + " requires exactly 2 arguments, got " + args.length);
		}			
		String value=args[0].stringValue();
		String value2=args[0].stringValue();
		return valueFactory.createLiteral(value.equals(value2));
	}

}
