src/test/de/hsmainz/cs/semgis/arqextension/test/geometry/relation/RelateMatchTest.java
package de.hsmainz.cs.semgis.arqextension.test.geometry.relation;

import static org.junit.Assert.assertEquals;

import org.eclipse.rdf4j.model.Value;
import org.eclipse.rdf4j.model.ValueFactory;
import org.eclipse.rdf4j.model.impl.SimpleValueFactory;
import org.eclipse.rdf4j.query.algebra.evaluation.function.postgis.geometry.relation.RelateMatch;
import org.junit.jupiter.api.Test;

public class RelateMatchTest {

	public static final String notIsocelesTriangle="POLYGON((1 2, 3 4, 5 6, 1 2))";
	
	@Test
	public void testRelateMatchTrue() {
		RelateMatch instance=new RelateMatch();
		ValueFactory valfac=SimpleValueFactory.getInstance();
		Value geo=valfac.createLiteral("212101212");
		Value geo2=valfac.createLiteral("212101212");
		Value result=instance.evaluate(valfac, geo,geo2);
		Value expResult=valfac.createLiteral(true);
		assertEquals(expResult, result);
	}
	
	@Test
	public void testRelateMatchFalse() {
		RelateMatch instance=new RelateMatch();
		ValueFactory valfac=SimpleValueFactory.getInstance();
		Value geo=valfac.createLiteral("212FF1212");
		Value geo2=valfac.createLiteral("212101212");
		Value result=instance.evaluate(valfac, geo,geo2);
		Value expResult=valfac.createLiteral(false);
		assertEquals(expResult, result);
	}
	
}
